# api/openapi.yaml
openapi: 3.0.3
info:
  title: sh rest api for ibm WA
  description: Thsi is a simple api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: https://psh-app1.12985tlgl6hs.eu-gb.codeengine.appdomain.cloud
  description: code engine
paths:
  /customer:
    summary: customer operation
    description: customer
    post:
      tags:
      - GetCustomeByPhone
      summary: Get ustomer by pthone
      description: Get ustomer by pthone
      operationId: CustomeByPthon
      requestBody:
        description: Update an existent pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustByPthoneRequest'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustByPthoneResponse'
        "405":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-swagger-router-controller: GetCustomeByPhone
  /customer/{cif}:
    summary: customer operation
    description: get customer by cif
    get:
      tags:
      - GetCustomeByCif
      summary: Get customer by cif
      description: Get ustomer by cif
      operationId: CustomeByCif
      parameters:
      - name: cif
        in: path
        description: cif of customer that needs to be fetched
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustByPthoneResponse'
        "405":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-swagger-router-controller: GetCustomeByCif
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        petId:
          type: integer
          format: int64
          example: 198772
        quantity:
          type: integer
          format: int32
          example: 7
        shipDate:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        description:
          type: string
    CustByPthoneRequest:
      type: object
      properties:
        phone:
          type: string
    CustByPthoneResponse:
      type: object
      properties:
        cif:
          type: string
          description: Customer ID
          example: "100012345"
        name:
          type: string
          description: customer short name
          example: Ercule Puaro
        phone:
          type: string
          description: Customer phone
          example: 380-50-222-33-44
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
      example:
        cif: "100012345"
        phone: 380-50-222-33-44
        name: Ercule Puaro
        accounts:
        - currency: currency
          account: account
        - currency: currency
          account: account
    Account:
      type: object
      properties:
        account:
          type: string
        currency:
          type: string
      example:
        currency: currency
        account: account